删除前判断是否存在

+++ handleGltf/split/process.js
@@ -14,10 +14,17 @@ function getInf(){
 }
 function deleteFiles() {
     var path="../../../_DATA_/model_sim/"
+    var haveList=[]
+    fs.readdirSync(path).forEach(function (s) {
+        haveList.push(s)
+    });
+    console.log(haveList)
     for(var i=0;i<names.length;i++){
         console.log(names[i],inf[names[i]].url)
         if(names[i]!==inf[names[i]].url){
-            fs.unlinkSync(path  + names[i]);
+            if(haveList[names[i]])
+                fs.unlinkSync(path  + names[i]);
         }
     }
+    console.log("finish!")
 }
